test: optimize redundant suites for faster runtime
