+++ src/gateway/openresponses-parity.test.ts
@@ -12,7 +12,7 @@ let InputFileContentPartSchema: typeof import("./open-responses.schema.js").Inpu
 let ToolDefinitionSchema: typeof import("./open-responses.schema.js").ToolDefinitionSchema;
 let CreateResponseBodySchema: typeof import("./open-responses.schema.js").CreateResponseBodySchema;
 let OutputItemSchema: typeof import("./open-responses.schema.js").OutputItemSchema;
-let buildAgentPrompt: typeof import("./openresponses-http.js").buildAgentPrompt;
+let buildAgentPrompt: typeof import("./openresponses-prompt.js").buildAgentPrompt;
 
 describe("OpenResponses Feature Parity", () => {
   beforeAll(async () => {
@@ -23,7 +23,7 @@ describe("OpenResponses Feature Parity", () => {
       CreateResponseBodySchema,
       OutputItemSchema,
     } = await import("./open-responses.schema.js"));
-    ({ buildAgentPrompt } = await import("./openresponses-http.js"));
+    ({ buildAgentPrompt } = await import("./openresponses-prompt.js"));
   });
 
   describe("Schema Validation", () => {
